// src/generated/api/catalog/models/productType.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * DCTD-UTC Commerce API
 * Contract for storefront and admin applications
 * OpenAPI spec version: 1.0.0
 */

export type ProductType = (typeof ProductType)[keyof typeof ProductType];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ProductType = {
  STANDARD: 'STANDARD',
  BUNDLE: 'BUNDLE',
} as const;