// src/generated/api/inventory/models/stockAdjustmentType.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * DCTD-UTC Commerce API
 * Contract for storefront and admin applications
 * OpenAPI spec version: 1.0.0
 */

export type StockAdjustmentType = (typeof StockAdjustmentType)[keyof typeof StockAdjustmentType];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const StockAdjustmentType = {
  CORRECTION: 'CORRECTION',
  OPENING_BALANCE: 'OPENING_BALANCE',
  MANUAL_RECEIPT: 'MANUAL_RECEIPT',
} as const;