// src/generated/api/inventory/models/stockAdjustmentReason.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * DCTD-UTC Commerce API
 * Contract for storefront and admin applications
 * OpenAPI spec version: 1.0.0
 */

export type StockAdjustmentReason =
  (typeof StockAdjustmentReason)[keyof typeof StockAdjustmentReason];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const StockAdjustmentReason = {
  MANUAL: 'MANUAL',
  COUNT_CORRECTION: 'COUNT_CORRECTION',
  INITIAL_STOCK: 'INITIAL_STOCK',
  EXTERNAL_RECEIPT: 'EXTERNAL_RECEIPT',
} as const;